refactor(usr): own explicit roots in contracts
